[ALL] Updating templates
