+++ .idea/fileTemplates/Compose Feature.kt.child.0.kt
@@ -1,38 +1,37 @@
 package ${PACKAGE_NAME}
 
-import kotlin.random.Random
+import com.cramsan.framework.core.compose.BaseViewModel
+import com.cramsan.framework.core.compose.ViewModelDependencies
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.SharedFlow
+import kotlinx.coroutines.flow.StateFlow
 
 /**
- * Events that can be triggered within the domain of the ${NAME} feature.
- *
- * Events are triggered from a ViewModel and are consumed by the UI.
- *
- */
-sealed class ${NAME}Event {
+ * ViewModel for the ${NAME} screen.
+ **/
+class ${NAME}ViewModel(
+    dependencies: ViewModelDependencies,
+) : BaseViewModel(dependencies) {
 
-    /**
-     * No operation.
-     */
-    data object Noop : ${NAME}Event()
-
-    /**
-     * Trigger application event. This event is sent to the application's view model to be handled.
-     */
-    data class TriggerApplicationEvent(
-        // Update this with the respective ApplicationEvent type.
-        val applicationEvent: ApplicationEvent,
-        val id: Int = Random.nextInt(),
-    ) : ${NAME}Event()
+    private val _uiState = MutableStateFlow(${NAME}UIState(
+        content = ${NAME}UIModel(""),
+        isLoading = false,
+    ))
     
     /**
-     * Trigger activity event. This event is triggered within a feature ViewModel and it will be
-     * consumed by the activity that is hosting this feature.
-     *
-     * This event is optional, it can be removed if not needed.
+     * UI state of the screen.
      */
-    data class TriggerActivityEvent(
-        // Update this with the respective ActivityEvent type.
-        val activityEvent: ActivityEvent,
-        val id: Int = Random.nextInt(),
-    ) : ${NAME}Event()
+    val uiState: StateFlow<${NAME}UIState> = _uiState
+
+    private val _event = MutableSharedFlow<${NAME}Event>()
+
+    /**
+     * Event flow to be observed.
+     */
+    val event: SharedFlow<${NAME}Event> = _event
+    
+    companion object {
+        private const val TAG = "${NAME}ViewModel"
+    }
 }
