+++ .idea/fileTemplates/Compose Feature.kt.child.1.kt
@@ -1,21 +1,17 @@
 package ${PACKAGE_NAME}
 
-import androidx.lifecycle.ViewModel
-import com.codehavenx.alpaca.frontend.appcore.managers.WorkContext
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.SharedFlow
-import kotlinx.coroutines.flow.StateFlow
+import androidx.compose.desktop.ui.tooling.preview.Preview
+import androidx.compose.runtime.Composable
 
-class ${NAME}ViewModel(
-    private val workContext: WorkContext,
-) : ViewModel() {
-    private val _uiState = MutableStateFlow(${NAME}UIState(
+// Move this file to the desktop source set. Once done, remove this line
+/**
+ * Preview for the ${NAME} feature screen.
+ */
+@Preview
+@Composable
+private fun ${NAME}ScreenPreview() {
+    ${NAME}Content(
         content = ${NAME}UIModel(""),
-        isLoading = false,
-    ))
-    val uiState: StateFlow<${NAME}UIState> = _uiState
-
-    private val _event = MutableSharedFlow<${NAME}Event>()
-    val event: SharedFlow<${NAME}Event> = _event
+        loading = false,
+    )
 }
