+++ .idea/fileTemplates/Compose Feature.kt.child.2.kt
@@ -1,19 +1,12 @@
 package ${PACKAGE_NAME}
 
 /**
- * ${NAME} UI model. You can use one or multiple models to represent
- * specific sections and component of the UI.
+ * UI state of the ${NAME} feature.
+ *
+ * This class models the top level state of the page.
+ * For modeling more specific details of the page, see the respective UI model class.
  */
-data class ${NAME}UIModel(
-    val name: String,
+data class ${NAME}UIState(
+    val content: ${NAME}UIModel,
+    val isLoading: Boolean,
 )
-
-/**
- * Converts the given [Any] object to an [${NAME}UIModel] object.
- */
-fun Any.to${NAME}UIModel(): ${NAME}UIModel {
-    // This is just a placeholder implementation. Replace the Any type and implement the conversion logic.
-    return ${NAME}UIModel(
-        name = TODO(),
-    )
-}
